Simplify array/single-value polymorphing
I ran into an issue with the `track_script` and `virtual_ipaddress`
settings when passing a single value rather than an array. Wrapping with
`Array()` fixed the issue for me. I decided to go through and use the
same strategy in a couple other places to simplify and limber up the
code a bit.

Specs passing locally on 1.8.7-p375, 2.0.0-p353, and 2.1.1