Add no_weight parameter on vrrp::script class
Despite what the keepalived documentation says default weight for a
script is not 2, if none is specified.  What happens instead is that if
no weight is specified, then the associated VRRP instances go into FAULT
state and go offline.  If noprempt is enabled, this appears to be the
only way to get the VRRP instance to track the state of the script.

This adds a new parameter to the keepalived::vrrp::script define to
allow specifying that the weight parameter should not be put in the
config file at all.  This is implemented this way, because many configs
probably don't have a weight specified because the class will provide a
weight config of '2' if none is provided.  Most people are probably
using this default.