Validate VRRP instance priority
Validate that the `priorty` param provided to `keepalived::vrrp::instance`
is semantically valid. This will catch errors early, before keepalived is
incorrectly configured.

The RFC states that the priority is an 8bit unsigned integer which means
that it must have a range of 0-255 inclusive. However it also goes on to say
that 0 and 255 are reserved for special operational use.

https://tools.ietf.org/html/rfc3768#section-5.3.4