Add tests for collect_unicast_peers
I also replaced a lot of explicit type validation by specifying types at
init, as recommended by bastelfreak. The unicast_source_ip parameter
previously was not required to be a valid IP address, so this broke an
existing test that was setting it to `_VALUE_`.

I also modified the test around setting explicit unicast_peers, since
this feature replaces a single concat fragment with several.