service: hasrestart/hasstatus: use Puppet defaults
This is a followup of 86b83d3bb245d212a14afa8a4ae9bf31db50761d. We don't
need to explicitly set hasrestart to false/true. All systems run systemd
and the units have a working restart option so we can use it. And
instead of hardcoding this we can just set it to undef so puppet figures
this out on its own.