Make config file validation optional
from time to time, keepalived deprecates config options. Those options
still work, but the config check will raise a warning and exit with 1.
The validation within puppet will fail. This can be problematic when you
are rolling out a new change and an already existing option will trigger
the warning. That will make it quite hard to roll out changes on older
systems.

With the new option the validation is opt-out and can be disabled for
older systems.